+++ Sources/Common/EitherSingleOrList.swift
