+++ Sources/Common/IndexQueryState.swift
