+++ Sources/Extensions/Optional+String.swift
