+++ Sources/Highlighting/HighlightedString.swift
