+++ Sources/QueryBuilder/QueryBuilder+HierarchicalFaceting.swift
