+++ Sources/Results/SearchResults.swift
@@ -310,22 +310,3 @@ extension SearchResults {
   
 }
 
-public struct MultiSearchResults: Codable {
-  
-  public let searchResults: [SearchResults]
-  
-  enum CodingKeys: String, CodingKey {
-    case results
-  }
-  
-  public init(from decoder: Decoder) throws {
-    let container = try decoder.container(keyedBy: CodingKeys.self)
-    searchResults = try container.decode([SearchResults].self, forKey: .results)
-  }
-  
-  public func encode(to encoder: Encoder) throws {
-    var container = encoder.container(keyedBy: CodingKeys.self)
-    try container.encode(searchResults, forKey: .results)
-  }
-  
-}
